backend/requirements.txt
Django==5.0.6
python-decouple==3.8
djangorestframework==3.15.1
django-cors-headers==4.3.1
openpyxl==3.1.5
python-dateutil==2.9.0.post0
